==> 01_ncis_nlp_data_scraping_cleaning.ipynb <==
## The CBS Secret Sauce Investigation: NCIS:NLP -- Scraping & Cleaning

#### Creator: Mitch Brinkman

from bs4 import BeautifulSoup
import requests

import pandas as pd
import numpy as np
import datetime as dt
import matplotlib.pyplot as plt
import seaborn as sns
import statsmodels.formula.api as smf
import patsy
import re
import pickle
import string
from sklearn.feature_extraction.text import CountVectorizer
from nltk.tokenize import word_tokenize, TreebankWordTokenizer
from nltk.stem import PorterStemmer, WordNetLemmatizer 
from nltk.util import ngrams
from nltk.corpus import stopwords
import nltk
nltk.download('punkt')
nltk.download('wordnet')

from sklearn.feature_extraction import text 
from sklearn.feature_extraction.text import CountVectorizer

import sys
sys.setrecursionlimit(20000)

# from cbs_func import NLPProcessor

from cbs_func import make_ep_list
from cbs_func import clean_tv_scripts
from cbs_func import ncis_data
from cbs_func import shield_data
from cbs_func import clean_sentiment

### Web Scraping

url_list = ['https://subslikescript.com/series/NCIS-364845','https://subslikescript.com/series/The_Shield-286486']

#### NCIS scripts

ncis_list = make_ep_list(url_list[0])

for i, num in enumerate(ncis_list):
    ncis_list[i] = num.replace('/series/NCIS-364845/','')

ncis_df = ncis_data(ncis_list)

ncis_df.set_index('ep_id',inplace=True)
ncis_df = ncis_df.astype(str)

pickle.dump(ncis_df, open("./data/raw/ncis_series.pkl", "wb"))

#### The Shield scripts

shield_list = make_ep_list(url_list[1])

for i, num in enumerate(shield_list):
    shield_list[i] = num.replace('/series/The_Shield-286486/','')

shield_df = shield_data(shield_list)

shield_df.set_index('ep_id',inplace=True)
shield_df = shield_df.astype(str)

pickle.dump(shield_df, open("./data/raw/shield_series.pkl", "wb"))

### Cleaning

add_stop_words = ['tony','abby','dinozzo','uh','um']
stop_words = text.ENGLISH_STOP_WORDS.union(add_stop_words)

def clean_tv_scripts(text):
    '''Make text lowercase, remove text in square brackets, 
    remove punctuation and remove words containing numbers.
    '''
    text = re.sub('([A-Z]{2,})', ' ', text)
    text = text.lower()
    text = re.sub('\[.*?\]', ' ', text)
    text = re.sub('\<.*?>', ' ', text)
    text = re.sub('\(.*?\)', ' ', text)
    text = re.sub('[%s]' % re.escape(string.punctuation), '', text)
    text = re.sub('\w*\d\w*', '', text)
    text = re.sub('\n', ' ', text)
    text = re.sub('♪', '', text)
    text = re.sub('  ', ' ', text)
    text = re.sub('   ', ' ', text)
    text = re.sub('    ', ' ', text)
    return text

big_wash = lambda x: clean_tv_scripts(x)

#### Corpus Prep for Sentiment Analysis

# # Pickle reload of raw file, for easy drop-in to notebook

ncis_df = pd.read_pickle('./data/raw/ncis_series.pkl')
shield_df = pd.read_pickle('./data/raw/shield_series.pkl')

#index reset and str conversion if necessary

# ncis_df.set_index('ep_id',inplace=True)
# shield_df.set_index('ep_id',inplace=True)
# ncis_df = ncis_df.astype(str)
# shield_df = shield_df.astype(str)

# creating a simple clean function referencing sentiment clean process
first_wash = lambda x: clean_sentiment(x)

clean_ncis = pd.DataFrame(ncis_df.dialogue.apply(first_wash))
clean_shield = pd.DataFrame(shield_df.dialogue.apply(first_wash))

pickle.dump(clean_shield, open("./data/processed/shield_sentiment.pkl", "wb"))
pickle.dump(clean_ncis, open("./data/processed/ncis_sentiment.pkl", "wb"))

#### NLP Pipeline Processing

# # Pickle reload of raw file, for easy drop-in to notebook

ncis_df = pd.read_pickle('./data/raw/ncis_series.pkl')
shield_df = pd.read_pickle('./data/raw/shield_series.pkl')

ncis_df.head()

class NLPProcessor:
    
    def __init__(self, vectorizer_class, tokenizer_function, cleaning_function, lemmer_function):
        self.vectorizer = vectorizer_class
        self.tokenizer = tokenizer_function
        self.cleaning_function = cleaning_function
        self.lemmer = lemmer_function
    
    def fit(self, corpus_list_to_fit):
        cleaned_corpus = list(map(self.cleaning_function, corpus_list_to_fit))
#         print(cleaned_corpus)
        tokenized_list = list(map(self.tokenizer, cleaned_corpus))
#         print(tokenized_list)
        lemmed_list = [' '.join(list(map(self.lemmer, item))) for item in tokenized_list]
#         print(lemmed_list)
        return self.vectorizer.fit(lemmed_list)
    
    def transform(self, corpus_list_to_clean):
        cleaned_corpus = list(map(self.cleaning_function, corpus_list_to_clean))
        tokenized_list = list(map(self.tokenizer, cleaned_corpus))
        lemmed_list = [' '.join(list(map(self.lemmer, item))) for item in tokenized_list]
        return pd.DataFrame(self.vectorizer.transform(lemmed_list).toarray(), 
                            columns=self.vectorizer.get_feature_names())

pd.DataFrame(doc_word.toarray(), index=ex_label, columns=vectorizer.get_feature_names()).head(10)

##### NCIS Pipeline

nlp = NLPProcessor(CountVectorizer(stop_words='english'), 
               TreebankWordTokenizer().tokenize, 
               clean_tv_scripts, 
               WordNetLemmatizer().lemmatize)

nlp.fit(ncis_df['dialogue'])
ncis_dtm = nlp.transform(ncis_df['dialogue'])

ncis_big_vocab = nlp.vectorizer
pickle.dump(ncis_big_vocab, open("ncis_cv.pkl", "wb"))

nlp = NLPProcessor(CountVectorizer(stop_words=stop_words,max_df=.95,
                                  min_df=.05), 
               TreebankWordTokenizer().tokenize, 
               clean_tv_scripts, 
               WordNetLemmatizer().lemmatize)

nlp.fit(ncis_df['dialogue'])
ncis_dtm_stopmm = nlp.transform(ncis_df['dialogue'])

ncis_vocab = nlp.vectorizer
pickle.dump(ncis_vocab, open("ncis_cv_stopmm.pkl", "wb"))

##### The Shield Pipeline

nlp = NLPProcessor(CountVectorizer(stop_words='english'), 
               TreebankWordTokenizer().tokenize, 
               clean_tv_scripts, 
               WordNetLemmatizer().lemmatize)

nlp.fit(shield_df['dialogue'])
shield_dtm = nlp.transform(shield_df['dialogue'])

shield_big_vocab = nlp.vectorizer
pickle.dump(shield_big_vocab, open("./data/cross_vec/shield_cv.pkl", "wb"))

nlp = NLPProcessor(CountVectorizer(stop_words='english',max_df=.95,min_df=.05), 
               TreebankWordTokenizer().tokenize, 
               clean_tv_scripts, 
               WordNetLemmatizer().lemmatize)

nlp.fit(shield_df['dialogue'])
shield_dtm_stopmm = nlp.transform(shield_df['dialogue'])

shield_vocab = nlp.vectorizer
pickle.dump(shield_vocab, open("./data/cross_vec/shield_cv_stopmm.pkl", "wb"))

##### Pickling & Indexing Each DTM for Modeling

ncis_dtm.index = ncis_df.index
ncis_dtm_stopmm.index = ncis_df.index
shield_dtm.index = shield_df.index
shield_dtm_stopmm.index = shield_df.index

pickle.dump(ncis_dtm, open("./data/dtm/ncis_dtm.pkl", "wb"))
pickle.dump(ncis_dtm_stopmm, open("./data/dtm/ncis_dtm_stopmm.pkl", "wb"))
pickle.dump(shield_dtm, open("./data/dtm/shield_dtm.pkl", "wb"))
pickle.dump(shield_dtm_stopmm, open("./data/dtm/shield_dtm_stopmm.pkl", "wb"))
