==> 03_ncis_nlp_sentiment_analysis.ipynb <==
## The CBS Secret Sauce Investigation: NCIS:NLP -- Sentiment Analysis

#### Creator: Mitch Brinkman

import pandas as pd
import numpy as np
import datetime as dt
import matplotlib.pyplot as plt
import seaborn as sns
import statsmodels.formula.api as smf
import patsy
import re
import pickle
import string
from sklearn.feature_extraction.text import CountVectorizer

ncis = pd.read_pickle('./data/processed/ncis_sentiment.pkl')
shield = pd.read_pickle('./data/processed/shield_sentiment.pkl')

ncis.tail(100)

#### The Shield Sentiment

# Create quick lambda functions to find the polarity and subjectivity of each routine
# Terminal / Anaconda Navigator: conda install -c conda-forge textblob
from textblob import TextBlob

pol = lambda x: TextBlob(x).sentiment.polarity


shield['polarity'] = shield['dialogue'].apply(pol)

shield['polarity'].mean()

shield.head()

shield.reset_index(inplace=True)

shield[['sea','#']] = shield.ep_id.str.split('/',expand=True)

shield_sea_sent = shield.groupby(['sea'],as_index=False,sort=False)['polarity'].mean()
shield_sea_sent

# Let's plot the results
import matplotlib.pyplot as plt

plt.rcParams['figure.figsize'] = [18, 14]
sns.set(font_scale = 1.4)
for index, dialogue in enumerate(shield.index):
    x = shield_sea_sent.sea
    y = shield_sea_sent.polarity
    plt.plot(x, y, color='red')
#     plt.text(x+.001, y+.001, ncis['episode'][index], fontsize=10)
    plt.ylim(.05, .15) 
    
# plt.figure(figsize=(15,12))
plt.xticks(rotation=20)
plt.title('The Shield: Season Sentiment Analysis', fontsize=20)
plt.xlabel('<-- Season 1 -------- Season 7 -->', fontsize=15)
plt.ylabel('<-- Negative -------- Positive -->', fontsize=15)

plt.savefig('shield_sentiment.png');
plt.show()

# Let's plot the results
import matplotlib.pyplot as plt

plt.rcParams['figure.figsize'] = [18, 14]
sns.set(font_scale = 1.4)
for index, dialogue in enumerate(shield.index):
    x = shield.ep_id
    y = shield.polarity
    plt.plot(x, y, color='red')
#     plt.text(x+.001, y+.001, ncis['episode'][index], fontsize=10)
#     plt.xlim(-.01, .12) 
    
# plt.figure(figsize=(15,12))
plt.title('The Shield: Per Episode Sentiment Analysis', fontsize=20)
plt.xlabel('<-- Season 1 -------- Season 7 -->', fontsize=15)
plt.ylabel('<-- Negative -------- Positive -->', fontsize=15)

plt.savefig('shield_sentiment.png');
plt.show()

#### NCIS Sentiment

ncis['polarity'] = ncis['dialogue'].apply(pol)

ncis.head()

ncis.reset_index(inplace=True)

ncis[['sea','#']] = ncis.ep_id.str.split('/',expand=True)

# ncis_sea_sent.drop('season',axis=1,inplace=True)

ncis_sea_sent = ncis.groupby(['sea'],as_index=False,sort=False)['polarity'].mean()
ncis_sea_sent

# ncis_sea_sent.sort_values('season', axis=0, ascending=True, inplace=True)

ncis['polarity'].mean()

# Let's plot the results
import matplotlib.pyplot as plt

plt.rcParams['figure.figsize'] = [18, 14]
sns.set(font_scale = 1.4)
for index, dialogue in enumerate(ncis.index):
    x = ncis_sea_sent.sea
    y = ncis_sea_sent.polarity
    plt.plot(x, y, color='blue')
#     plt.text(x, y, ncis_sea_sent['sea'][index], fontsize=10)
    plt.ylim(.07, .12) 
    
plt.xticks(rotation=20)
plt.title('NCIS: Season by Season Sentiment', fontsize=20)
plt.xlabel('<-- Season 1 -------- Season 17 -->', fontsize=15)
plt.ylabel('<-- Negative -------- Positive -->', fontsize=15)

plt.savefig('ncis_season_sentiment.png');
plt.show()
