==> 04_ncis_nlp_topic_modeling.ipynb <==
# The CBS Secret Sauce Investigation: NCIS:NLP -- Topic Modeling

#### Creator: Mitch Brinkman

### Imports

import pandas as pd
import numpy as np
import datetime as dt
import matplotlib.pyplot as plt
import seaborn as sns
import statsmodels.formula.api as smf
import patsy
import re
import pickle
import string
from sklearn.feature_extraction.text import CountVectorizer
from nltk.tokenize import word_tokenize, TreebankWordTokenizer
from nltk.stem import PorterStemmer, WordNetLemmatizer 
from nltk.util import ngrams
from nltk import pos_tag
from gensim import matutils, models
import scipy.sparse
nltk.download('punkt')
nltk.download('wordnet')

## Topic Modeling - NCIS

# Let's read in our document-term matrix


ncis_data = pd.read_pickle('./data/dtm/ncis_dtm_stopmm.pkl')
ncis_full_data = pd.read_pickle('./data/dtm/ncis_dtm.pkl')
ncis_full_tdm = ncis_full_data.transpose()
ncis_tdm = ncis_data.transpose()
ncis_tdm.head()

# converting to gensim format

ncis_sparse_counts = scipy.sparse.csr_matrix(ncis_tdm)
ncis_corpus = matutils.Sparse2Corpus(ncis_sparse_counts)
ncis_cv = pickle.load(open("./data/cross_vec/ncis_cv_stopmm.pkl", "rb"))
id2word = dict((v, k) for k, v in ncis_cv.vocabulary_.items())

# Gensim also requires dictionary of the all terms and their respective location in the term-document matrix
# cv = pickle.load(open("shield_cv_stopmm.pkl", "rb"))
# id2word = dict((v, k) for k, v in cv.vocabulary_.items())

# Now that we have the corpus (term-document matrix) and id2word (dictionary of location: term),

ncis_lda = models.LdaModel(corpus=ncis_corpus, id2word=id2word, num_topics=4, passes=80)
ncis_lda.print_topics()

# LDA for num_topics = 3
ncis_lda = models.LdaModel(corpus=ncis_corpus, id2word=id2word, num_topics=8, passes=10)
ncis_lda.print_topics()

# LDA for num_topics = 4
ncis_lda = models.LdaModel(corpus=ncis_corpus, id2word=id2word, num_topics=5, passes=10)
ncis_lda.print_topics()

# converting to gensim format

ncis_sparse_counts = scipy.sparse.csr_matrix(ncis_full_tdm)
ncis_corpus = matutils.Sparse2Corpus(ncis_sparse_counts)
ncis_cv = pickle.load(open("./data/cross_vec/ncis_cv.pkl", "rb"))
id2word = dict((v, k) for k, v in ncis_cv.vocabulary_.items())

# Now that we have the corpus (term-document matrix) and id2word (dictionary of location: term),

ncis_lda = models.LdaModel(corpus=ncis_corpus, id2word=id2word, num_topics=4, passes=80)
ncis_lda.print_topics()

# LDA for num_topics = 3
ncis_lda = models.LdaModel(corpus=ncis_corpus, id2word=id2word, num_topics=8, passes=10)
ncis_lda.print_topics()

# LDA for num_topics = 4
ncis_lda = models.LdaModel(corpus=ncis_corpus, id2word=id2word, num_topics=5, passes=10)
ncis_lda.print_topics()

## Topic Modeling - The Shield

# Let's read in our document-term matrix


shield_data = pd.read_pickle('./data/dtm/shield_dtm_stopmm.pkl')
shield_full_data = pd.read_pickle('./data/dtm/shield_dtm.pkl')
shield_tdm = shield_data.transpose()
shield_full_tdm = shield_full_data.transpose()
shield_tdm.head()

# converting to gensim format
sparse_counts = scipy.sparse.csr_matrix(shield_tdm)
corpus = matutils.Sparse2Corpus(sparse_counts)
cv = pickle.load(open("./data/cross_vec/shield_cv_stopmm.pkl", "rb"))
id2word = dict((v, k) for k, v in cv.vocabulary_.items())

# LDA modeling w/ different numbers of topics & passes

shield_lda = models.LdaModel(corpus=corpus, id2word=id2word, num_topics=5, passes=80)
shield_lda.print_topics()

# LDA for num_topics = 10
shield_lda = models.LdaModel(corpus=corpus, id2word=id2word, num_topics=10, passes=10)
shield_lda.print_topics()

# LDA for num_topics = 4
shield_lda = models.LdaModel(corpus=corpus, id2word=id2word, num_topics=4, passes=10)
lda.print_topics()

# converting to gensim format
sparse_counts = scipy.sparse.csr_matrix(shield__full_tdm)
corpus = matutils.Sparse2Corpus(sparse_counts)
cv = pickle.load(open("./data/cross_vec/shield_big_vec_lib.pkl", "rb"))
id2word = dict((v, k) for k, v in cv.vocabulary_.items())

# LDA modeling w/ different numbers of topics & passes

shield_lda = models.LdaModel(corpus=corpus, id2word=id2word, num_topics=5, passes=80)
shield_lda.print_topics()

# LDA for num_topics = 10
shield_lda = models.LdaModel(corpus=corpus, id2word=id2word, num_topics=10, passes=10)
shield_lda.print_topics()

# LDA for num_topics = 4
shield_lda = models.LdaModel(corpus=corpus, id2word=id2word, num_topics=4, passes=10)
lda.print_topics()

## LEFTOVERS - scratch area

# # Read in the cleaned data, before the CountVectorizer step
# data_clean = pd.read_pickle('data_clean.pkl')
# data_clean

# # Apply the nouns function to the transcripts to filter only on nouns
# data_nouns = pd.DataFrame(data_clean.transcript.apply(nouns))
# data_nouns

# # Create a new document-term matrix using only nouns
# from sklearn.feature_extraction import text
# from sklearn.feature_extraction.text import CountVectorizer

# # Re-add the additional stop words since we are recreating the document-term matrix
# add_stop_words = ['like', 'im', 'know', 'just', 'dont', 'thats', 'right', 'people',
#                   'youre', 'got', 'gonna', 'time', 'think', 'yeah', 'said']
# stop_words = text.ENGLISH_STOP_WORDS.union(add_stop_words)

# # Recreate a document-term matrix with only nouns
# cvn = CountVectorizer(stop_words=stop_words)
# data_cvn = cvn.fit_transform(data_nouns.transcript)
# data_dtmn = pd.DataFrame(data_cvn.toarray(), columns=cvn.get_feature_names())
# data_dtmn.index = data_nouns.index
# data_dtmn

# # Create the gensim corpus
# corpusn = matutils.Sparse2Corpus(scipy.sparse.csr_matrix(data_dtmn.transpose()))

# # Create the vocabulary dictionary
# id2wordn = dict((v, k) for k, v in cvn.vocabulary_.items())

# # Let's start with 2 topics
# ldan = models.LdaModel(corpus=corpusn, num_topics=2, id2word=id2wordn, passes=10)
# ldan.print_topics()

# # Let's try topics = 3
# ldan = models.LdaModel(corpus=corpusn, num_topics=3, id2word=id2wordn, passes=10)
# ldan.print_topics()

# # Let's try 4 topics
# ldan = models.LdaModel(corpus=corpusn, num_topics=4, id2word=id2wordn, passes=10)
# ldan.print_topics()

# shield_clean = pd.read_pickle('./pickles/clean_shield_series.pkl')

# # Let's create a function to pull out nouns from a string of text
# def nouns_adj(text):
#     '''Given a string of text, tokenize the text and pull out only the nouns and adjectives.'''
#     is_noun_adj = lambda pos: pos[:2] == 'NN' or pos[:2] == 'JJ'
#     tokenized = word_tokenize(text)
#     nouns_adj = [word for (word, pos) in pos_tag(tokenized) if is_noun_adj(pos)] 
#     return ' '.join(nouns_adj)

# # Apply the nouns function to the transcripts to filter only on nouns
# data_nouns_adj = pd.DataFrame(shield_clean.dialogue.apply(nouns_adj))
# data_nouns_adj

# # Create a new document-term matrix using only nouns and adjectives, also remove common words with max_df
# cvna = CountVectorizer(stop_words='english', max_df=.95)
# data_cvna = cvna.fit_transform(data_nouns_adj.dialogue)
# data_dtmna = pd.DataFrame(data_cvna.toarray(), columns=cvna.get_feature_names())
# data_dtmna.index = data_nouns_adj.index
# data_dtmna

# # Create the gensim corpus
# corpusna = matutils.Sparse2Corpus(scipy.sparse.csr_matrix(data_dtmna.transpose()))

# # Create the vocabulary dictionary
# id2wordna = dict((v, k) for k, v in cvna.vocabulary_.items())

# # Let's start with 2 topics
# ldana = models.LdaModel(corpus=corpusna, num_topics=5, id2word=id2wordna, passes=10)
# ldana.print_topics()
